requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.6.0
pydantic==2.9.2
dnspython==2.4.2
email-validator==2.1.0.post1
mypy==1.13.0
ruff==0.0.47
passlib[bcrypt]>=1.7.4
jinja2>=3.1.4
httpx>=0.25.1
bcrypt==4.0.1
pyjwt==2.9.0
motor>=3.6